cleaning up

+++ backend/models.py
@@ -19,8 +19,8 @@ class Area(models.Model):
 
 class Piece(models.Model):
     name = models.CharField(max_length=30)
-    min_size = models.IntegerField(default=0)
-    max_size = models.IntegerField(default=0)
+    min_size = models.FloatField(default=0)
+    max_size = models.FloatField(default=0)
     SLCD = models.BooleanField(default=False)
     offset = models.BooleanField(default=False)
     gear_type = models.CharField(blank=True, null=True, max_length=50)
